feat: LabelAdd - 레이블 생성 시 속성 정보 구현
- Label 랜덤 색상
- Label text 두가지 색상  toggle
- ghojeong/issue-tracker/#134
