+++ frontend/issue-cracker/src/components/layout/LabelList/LabelAdd/LabelBox.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import styled from 'styled-components';
+import { useRecoilValue } from 'recoil';
+import LabelSmallGroup from '../../../common/group/LabelSmallGroup';
+import { labelAddInputState } from '../../../../store/Recoil';
+import { LABEL as L } from '../../../../utils/const';
+//labelAdd
+const LabelBox = () => {
+  const labelAdd = useRecoilValue(labelAddInputState);
+  return (
+    <LabelBoxStyle>
+      <LabelSmallGroup
+        label={labelAdd.title === '' ? L.NAME : labelAdd.title}
+        backgroundColor={labelAdd.backgroundColor}
+        color={labelAdd.textColor}
+      />
+    </LabelBoxStyle>
+  );
+};
+export default LabelBox;
+
+const LabelBoxStyle = styled.div``;
